[Easy] Check additional reward target column (#414)
This PR adds a check for a column to the reward target dataframe.

The code had previously changed to require a `"pool_address"` column.
This was reflected in the test `test_prepare_transfers`, as the column
is required there. The column was not, however, checked in
`validate_df_columns`. This PR adds that check and also suitably adapts
other tests to populate the `"pool_address"` column.

This should not lead to a behaviour change of the code.