[Easy] Remove option to consolidate transfers (#416)
This PR removes a command line argument to merge transfers and all code
associated to it.

With our current code, switching on this option would result in
transfers merging transfers which have not been merged on purpose.
Examples for this are transfers for quote and batch rewards, as well as
transfers for protocol fees and partner fees. Removing the option makes
it more difficult to use the script in a wrong way.

At the end of the restructuring of the solver accounting, we might also
want to merge transfers again. This will, however, not be compatible
with the simple logic currently implemented. Thus, keeping the current
code has little benefits.

Removing the command line option allows to also remove the associated
code and tests. This makes the rest of the code easier to change.

This change is a breaking change of the command line API. We currently
do not use the option anywhere, so the change should not affect current
operations.