[Easy] Small rework of transfer_file initialization (#415)
This PR changes the initialization step in the transfer_file script. It
does not change the behavior of the code.

The main change is to move the initialization of the dune fetcher into
the main script and out of the parsing of arguments. This will make it
easier to later explicitly pass config into initialization steps. It is
also easier for me to understand, what individual steps in the script
do. This change required adding a start argument to `ScriptArgs`.

I also reordered the script a bit and to make it easier to consistently
pass config objects as last argument of function calls.