Revert change about filtering dataframes (#417)
This PR reverts a change introduces in #407. The PR introduced a
somewhat unrelated change to filter duplicate columns from data frames
before merging.

In local testing it seems that the data frames do not actually contain
such columns. Thus, the somewhat difficult to read code fragment does
not have any effect.

If the code is indeed required, I would suggest restructuring it for
clarity.

If it is important that this column only appears in one data frame, one
could add a check to `validate_df_columns`.

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>