Restructure price fetching (#421)
This PR slightly restructures the price fetching for converting token
amount for rewards.

- Instead of having functions for converting to and from ETH
(`eth_in_token`, `token_in_eth`) we just provide an exchange rate for
conversion between two tokens (`exchange_rate_atoms`). This will make it
easier to add conversion between COW and the native token of the chain,
as well as between XDAI and ETH.
- Instead of using token ids as argument, the main function accepts
addresses. This is mostly to avoid leaking the implementation of the
price fetching (token id on coin paprika) into the rest of the code. It
also avoids a circular dependency in making the reward token part of the
configuration in #412.

The `TokenId` abstraction still feels a bit stange as there is some
overlap with `Token` due to decimals. Supporting USDC seems a bit
artificial since that is only used for testing and checking that
decimals are correctly accounted for.

Tests for price fetching were adapted accordingly. Transfers for the
accounting week starting on 2024-10-29 only differ on the level of
floating point precision.

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>