Remove constants from payout datum (#418)
This PR makes the payout datum independent of constants.

This will make it easier to pass a config object through the code, as it
will not be needed to pass to the creation of transfers from a payout
datum.

- This PR adds an explicit separation between reward and buffer
accounting target. This way we do not need to check for the bonding pool
when creating transfers. This change requires adding a new column,
`"buffer_accounting_target"`, to the payment dataframe.
- The type of the service fee field is changed to `Fraction` from
`bool`. This makes it possible to avoid passing the service fee factor
to several functions related to creating transfers. This requires a
change in meaning of the `"service_fee"` column.
- The reward token address (of COW, which might depend on the network)
is explicitly part of the payment datum. This requires adding a column
`"reward_target_address"`, to the payment dataframe.

Tests are adapted accordingly. I have not yet tested the effect on
payments. In principle, the behavior should not change.