Fix service fee log (#426)
This PR fixes an issue with the log for service fees.

The meaning of the attribute `service_fee` of the
`RewardAndPenaltyDatum` CHANGED IN #418. This change was not correctly
handled in the computation of the total service fee. The reward was thus
multiplied by the square of the service fee instead of just the bare
service fee. This is fixed in the PR.

No tests are changed since the log is not tested. Maybe we need to add a
test for the property that `total_reward_cow() + reward_scaling() *
quote_reward_cow + total_service_fee() == primary_reward_cow +
quote_reward_cow`.