[Multi-Chain] Add config file (#412)
This PR adds a config file. This configuration is supposed to be
extended to other chains in future PRs.

This PR adds a config `AccountingConfig` which consists of multple
configurations for different aspects of the accounting.

The config is used as a global config. At the moment it is not
explicitly passed to any function.

Next steps:
- Pass config explicitly to functions.
- Make some of the code more general to allow for using it on other
chains.
- Add support for other networks. (At the moment I only copied the
testing setup on other chains from the original code.)
- Make config more consistent.

There are still hardcoded constants somewhere in the code. One place is
when fetching prices in converting token values, via an ID. The id
cannot be part of the config due to circular imports. Instead, an
address or id string should be passed and the pricing code should turn
it into an appropriate type.

### How to review
The first two commits (the names are wrong, I might force push correct
ones) add a config and use the config throughout the code. The third
commit just removes the old constants file.

One discussion should mostly be about the first commit, i.e. the one
adding the configuration classes in `config.py`.
- Should the config be structured differently?
- Do we need special treatment of secrets?

Another discussion would be around how to move forward from here.
- Should we pass the config struct more explicitly? (That is what I did
in the more complete code I have.)
- Which parameters are probably still missing for making it work on
other chains?

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>