Add batch and order data jobs (#448)
This PR adds some functionality that is disjoint from the rest of the
payouts script, and can be used to set up a cronjob that syncs some
batch and order data to the analytics db. A separate dune-sync job is
meant to be introduced in order to sync these tables to the
corresponding dune tables.

In order to avoid conflict with the current state of the script, the
batch and order data queries, which are almost identical to the order
rewards and batch rewards queries, are presented as separate files. The
block intervals on which the queries work are still to be
double-checked.

To trigger the scripts locally and test the functionality, one can run

`python -m src.data_sync.sync_data --sync-table order_data `

or 

`python -m src.data_sync.sync_data --sync-table batch_data `

---------

Co-authored-by: Felix Henneke <felix.henneke@protonmail.com>