Revert column name change to surplus_fee (#476)
PR #471 adjusted the sql queries to work with the renamed column in the
backend db; surplus_fee was changed to executed_fee. However, dune-sync
and some dune-queries break because of that so this PR just renames the
final column stored during the order data sync job to surplus_fee.

A proper migration to executed_fee needs to happen, but this can wait a
bit.