Fix dashboard url (#487)
This PR fixes the dashboard url for verifying payments.

It now takes into account the network and network specific parameters.

A follow up PR should fix parameters for unusual slippage.

At some point the function `dashboard_url` should be moved to some other
file. It is not related to creating transfer files. The old place with
the accounting period does not work though, as the url does depend on
additional parameters.

This closes #473.