No service fee on penalties (#424)
With this PR, if a solver receives a negative reward for an accounting
week, that amount is not multiplied by `reward_scaling`.

Before the change, ~negative~ rewards were multiplied by
`reward_scaling` independent of the sign of the reward. For positive
rewards, this does what it should, as a fraction of the reward is
withheld from the solver and becomes part of a bounty budget. For
negative rewards, the calculation implied that part of the penalty is
withheld and it reduces the bounty budget.

With the change, negative _batch_ rewards are not multiplied by
`reward_scaling`. Quote rewards are left as is.

The proposed implementation changes some code in 3 places. Ideally there
would only be one change required. Note, that quote rewards are always
non-negative and no change in the formula is required.

Alternatively, the total cow reward could be used to decide on the
scaling, instead of handling batch and quote rewards separately.

Tests have not been adapted yet.

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>