Restructure payouts file (#435)
This PR is an attempt at implementing #427. The PR does not change the
values of final results (up to rounding of floats). It changes the
structure of the code and intermediate representations.

As a first step, it separates the computation of different parts of the
accounting into different functions.

```python
    # compute individual components of payments
    solver_info = compute_solver_info(
        reward_target_df,
        service_fee_df,
        config,
    )
    rewards = compute_rewards(
        batch_data,
        quote_rewards_df,
        exchange_rate_native_to_cow,
        config.reward_config,
    )
    protocol_fees = compute_protocol_fees(batch_data)
    partner_fees = compute_partner_fees(batch_data, config.protocol_fee_config)
    buffer_accounting = compute_buffer_accounting(batch_data, slippage_df)
```
Those functions are implemented in separate files 

The results of these steps are converted into data frames for solver
payments and for protocol and partner fee payments.

```python
    # combine into solver payouts and partner payouts
    solver_payouts = compute_solver_payouts(
        solver_info, rewards, protocol_fees, buffer_accounting
    )
    partner_payouts = (
        partner_fees  # no additional computation required here at the moment
    )
```

Payout data on transfers and overdrafts is then computed from solver and
parner payouts data.

```python
    payouts = prepare_payouts(solver_payouts, partner_payouts, dune.period, config)
```

The code can be tested to produce the same transfer files as the old
code.
Tests have been adapted and cover essentially what was tested before.
There is a bit more strictness in the testing of the separate
computations of rewards, protocol fees, partner fees, etc.
There is no end-to-end test for payments yet. This should be added at
some point.

Future changes could remove data frames from intermediate results. This
would make it easier to have correct types and detect and handle missing
data. Data for the different parts of the accounting can be changed to
use intermediate tables generated by `src/data_sync/sync_data.py`.

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>