Add optional excluded auctions list (#516)
This PR allows the script to set the capped payment of an auction to
zero, if it is part of an "excluded auctions" list. Note that the
"correct", i.e., full, reward will still be synced to Dune, and a
similar list of excluded auctions will be needed on Dune as well.

Note that protocol/partner/network fees are still paid for such excluded
auctions, and slippage accounting is taking place for those as well.