Use analytics db for solver rewards (#547)
This PR implements simple fetching and parsing from the new analytics
db.

The code is run by adding the `--fetch-from-analytics-db` flag.

It is mostly for testing the old and new accounting in parallel. After
switching to the new accounting pipelin, most of the code in solver
rewards can be removed.

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>