Implement CIP-68 (#557)
This PR implements the changes proposed in CIP-68. The main difference
is that all COW transfers are proposed on mainnet, while buffer
accounting is still handled on individual chains.

With this PR, two transactions are proposed on every chain.
- One transaction containing COW transfers on mainnet.
- One transaction containing native transfers on the respective chain.

In case of a run without the `--post-tx` flag, two csv files are
created.

I have tested the creation of csv files and the bare creation of
transactions. I have not tested the actual proposal of transactions yet.